+++ backend/CadastroLivros/Data/Entities/Livro.cs
@@ -2,7 +2,7 @@
 {
     public class Livro
     {
-        public int CodL { get; set; }
+        public long CodL { get; set; }
         public string Titulo { get; set; }
         public string Editora { get; set; }
         public int Edicao { get; set; }
