+++ backend/CadastroLivros/Services/Interfaces/IAutorService.cs
@@ -4,10 +4,10 @@ namespace CadastroLivros.Services.Interfaces
 {
     public interface IAutorService
     {
-        Task<AutorDTO?> Read(int codAu);
+        Task<AutorDTO?> Read(long codAu);
         Task<AutorDTO> Insert(AutorDTO autor);
         Task Update(AutorDTO autor);
-        Task Delete(int codAu);
+        Task Delete(long codAu);
         Task<IEnumerable<AutorDTO>> ReadPage(int page = 0, int pageSize = 40);
     }
 }
